feat: Update car image URL in SportsCarRentalApplication
The car image URL in the SportsCarRentalApplication has been updated to a new URL. This change improves the accuracy of the displayed car image.

Refactor the Car class to return the car image URL as a string

This refactor updates the Car class to return the car image URL as a string. This improves the consistency and readability of the code.

Recent user commits:
- Comments deleted
- Cars added
- New Version
- refactor: Update Car class to return car image URL as a string
- Merge branch 'main' of https://github.com/AltPhilipp/Pizza_Reference
- New Entities
- image
- carimage
- Merge branch 'main' of https://github.com/AltPhilipp/MotorSportRentals
- added image